// src/modules/garmentsProduction/reports/embellishment/embellishment-delivery-report/embellishment-delivery-report-type.ts
export type EmbellishmentDeliveryReportType = {
  DELIVERY_CHALLAN_NO?: string;
  SUPPLIER?: string;
  SUPPLIER_ADDRESS?: string;
  DELIVERY_DATE: Date;
  WORK_ORDER_NO?: string;
  EMBELLISHMENT_ORDERNO?: string;
  PARTS?: string;
  BUYER?: string;
  STYLE?: string;
  PONO?: string;
  COLOR_NAME?: string;
  SIZE_NAME?: string;
  OS_BUYER?: string;
  OS_STYLE?: string;
  OS_PONO?: string;
  DELIVERY_QTY: number;
  PREPARED_BY?: string;
  EMB_TYPE?: string;
  REASON?: string;
  REJECT_QTY: number;
};